Update several files in shell/common.d to not load if on a Mac

+++ shell/common.d/sudo.sh
@@ -1,3 +1,5 @@
+[[ $(uname) == 'Darwin' ]] && return
+
 alias shutdown="sudo shutdown -h +0"
 alias reboot="sudo reboot"
 alias partitions2="sudo blkid"
